funcX-v6/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 


azure-functions
azure-ai-documentintelligence
azure-storage-blob
azure-cosmos
python-dotenv
PyMySQL
